move a new file out of sdkutils that shouldn't be there
