+++ Tasks/Common/awsConnectionParameters.ts
